+++ src/prelude/prelude.mli
@@ -1,3 +1,3 @@
 val prelude : string
 val internals : string
-val prim_module : timers:bool -> experimental_stable_memory:int -> string
+val prim_module : timers:bool -> string